refactor(theme): Create reusable theme switcher component
- Created a new reusable `ThemeSwitcher` component from the theme selection UI in the Settings page.
- Replaced the inconsistent theme toggle in the Overview page with the new `ThemeSwitcher` component.
- Refactored the Settings page to use the new `ThemeSwitcher` component.
- This change makes the theme selection UI consistent across the application and improves code maintainability.
