+++ pages/Settings.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import Header from '../components/Header';
 import { Theme } from '../types';
+import ThemeSwitcher from '../components/ThemeSwitcher';
 
 interface SettingsProps {
     theme: Theme;
@@ -8,12 +9,6 @@ interface SettingsProps {
 }
 
 const Settings: React.FC<SettingsProps> = ({ theme, setTheme }) => {
-    const themeOptions: { name: Theme, label: string }[] = [
-        { name: 'light', label: 'Light' },
-        { name: 'dark', label: 'Dark' },
-        { name: 'bw', label: 'B & W' }
-    ];
-
     return (
         <div className="p-6">
             <Header title="Settings" description="Manage your account and preferences." showPlatformSelector={false} />
@@ -46,22 +41,7 @@ const Settings: React.FC<SettingsProps> = ({ theme, setTheme }) => {
                             <p className="font-medium">Interface Theme</p>
                             <p className="text-sm text-on-surface-secondary">Select your preferred interface style.</p>
                         </div>
-                        <div className="flex items-center space-x-2 bg-background p-1 rounded-full">
-                            {themeOptions.map((option) => (
-                                <button
-                                    key={option.name}
-                                    onClick={() => setTheme(option.name)}
-                                    className={`px-4 py-1.5 text-sm font-medium rounded-full transition-colors duration-200 focus:outline-none focus:ring-2 focus:ring-primary focus:ring-offset-2 focus:ring-offset-background ${
-                                        theme === option.name
-                                            ? 'bg-primary text-white shadow-md'
-                                            : 'text-on-surface-secondary hover:text-on-surface'
-                                    }`}
-                                    aria-pressed={theme === option.name}
-                                >
-                                    {option.label}
-                                </button>
-                            ))}
-                        </div>
+                        <ThemeSwitcher theme={theme} setTheme={setTheme} />
                     </div>
                 </div>
